Use the type state pattern to manage building the pipeline
 - Use a Lisp-style HList to track which data stores are added during
   the peristence-adding stage, so that they can be made available
   during the load-adding stage.
 - Remove the runtime pipeline validation logic and errors.rs module

These changes are overkill from a design perspective but have been made 
as an exercise to use the type state pattern with some fairly advanced type 
system utilisation. The book offers two routes to store data but doesn't 
require them both to be made available.

In reality, the design would select a single data store for each specific 
bit of data, to maintain a single source of truth.